rg-190: * fix touch drag and add hook for sensors
