+++ frontend/src/bundles/common/hooks/hooks.ts
@@ -3,6 +3,7 @@ export { useAppForm } from './use-app-form/use-app-form.hook.js';
 export { useAppSelector } from './use-app-selector/use-app-selector.hook.js';
 export { useModal } from './use-modal/use-modal';
 export { useRedirect } from './use-redirect/use-redirect';
+export { useTemplateSensors } from './use-template-sensors/use-template-sensors.hook';
 export { useCallback, useEffect, useMemo, useState } from 'react';
 export { useController as useFormController } from 'react-hook-form';
 export { useLocation, useNavigate } from 'react-router-dom';
